fix: normalize script naming and comparator contract consistency
- Rename generate_rung_charts_v2.py to generate_rung_charts.py (match plan)
- Remove legacy generate_rung_charts.py (eval JSONL reader superseded by
  CSV-first chart generator)
- Update run_rung.py Step 7 to pass --tables-dir/--chart-data-dir args
  matching the CSV-first CLI
- Update all references in docs, tests, and orchestrator
- Add per-contract faceting support to generate_comparator_rankings() and
  generate_behavior_by_contract() — ready for bidders_by_contract data
  when extract_comparator_cis.py is updated (documented TODO)
- Add test for per-contract faceting with enriched fixture data
